https://github.com/Dopeboi029/Dopeboi029.git
GitHub is the developer company. We make it easier for
developers to be developers: to work together, to solve
challenging problems, and to create the world's most
important technologies. We foster a collaborative
community that can come together-as individuals and
in teams-to create the future of software and make a
difference in the world.
The GitHub platform is where developers can create,
share, and ship the best code possible. Our community
is made up of students, hobbyists, consultants,
enterprise professionals, partners and executives,
building software in the way that works best for them.
GitHub is much more than code. It is the home of open
source collaboration. It is where new developers get
started and where experienced developers expand their
knowledge. It is a community where developers come
together to solve the unsolvable and test the limits of
what software can do.